+++ app/(main)/products/[productId]/component/DescriptionTab.jsx
@@ -1,0 +1,37 @@
+const DescriptionTab = ({ product }) => {
+  return (
+    <div className="prose max-w-none pamber-amber font-bengali">
+      <h3>জামদানি শাড়ির ঐতিহ্য</h3>
+      <p>
+        বাংলাদেশের জামদানি শাড়ি এক অনন্য ঐতিহ্য, যা শতাব্দীর পর শতাব্দী ধরে
+        সূক্ষ্ম হাতের কাজের জন্য বিখ্যাত। এই {product.name} শাড়িটি পুরোপুরি
+        হাতে বোনা এবং ঐতিহ্যবাহী পদ্ধতিতে তৈরি করা হয়েছে।
+      </p>
+
+      <p>এই শাড়ির বিশেষ বৈশিষ্ট্য:</p>
+      <ul>
+        <li>১০০% খাঁটি সুতি দিয়ে বোনা</li>
+        <li>ঐতিহ্যবাহী {product.pattern} প্যাটার্ন</li>
+        <li>বিশেষ {product.color} রং সমন্বয়</li>
+        <li>অত্যন্ত সূক্ষ্ম হাতের কাজ</li>
+        <li>{product.region} অঞ্চলের দক্ষ তাঁতি দ্বারা নির্মিত</li>
+      </ul>
+
+      <p>
+        জামদানির সূক্ষ্ম বুনন কৌশল ২০১৩ সালে ইউনেস্কো দ্বারা অমূর্ত সাংস্কৃতিক
+        ঐতিহ্য হিসেবে স্বীকৃত হয়েছে। এই শাড়িটি বাংলাদেশের জাতীয় শিল্পকলার
+        একটি অনন্য নিদর্শন, যা পরিধান করে আপনি বাংলাদেশের ঐতিহ্য এবং সংস্কৃতির
+        অংশ হতে পারবেন।
+      </p>
+
+      <h3>ব্যবহার এবং যত্ন</h3>
+      <p>
+        জামদানি শাড়ি যত্ন সহকারে হাত দিয়ে ধুয়ে নিতে হবে। শাড়িটি ছায়ায়
+        শুকাতে দিন এবং কখনই সরাসরি সূর্যালোকে রাখবেন না। ইস্ত্রি করার সময় হালকা
+        গরম ইস্ত্রি ব্যবহার করুন।
+      </p>
+    </div>
+  );
+};
+
+export default DescriptionTab;
