+++ data/review.json
@@ -1,0 +1,24 @@
+[
+  {
+    "name": "রুমানা আক্তার",
+    "rating": 5,
+    "date": "৫ মার্চ, ২০২৩",
+    "title": "অসাধারণ কোয়ালিটি!",
+    "comment": "শাড়িটি অসাধারণ! কোয়ালিটি এবং ডিজাইন দুটোই চমৎকার। আমার বিয়ের অনুষ্ঠানে পরেছিলাম এবং সবাই শাড়ির প্রশংসা করেছিল।",
+    "images": ["/assets/review-1.jpg", "/assets/review-2.jpg"]
+  },
+  {
+    "name": "তানিয়া রহমান",
+    "rating": 4,
+    "date": "১২ ফেব্রুয়ারি, ২০২৩",
+    "title": "সুন্দর ডিজাইন",
+    "comment": "খুব সুন্দর জামদানি শাড়ি, তবে রঙটি ছবির চেয়ে একটু ভিন্ন। তবে মানের দিক দিয়ে খুবই ভালো।"
+  },
+  {
+    "name": "আসিফ আহমেদ",
+    "rating": 5,
+    "date": "২২ জানুয়ারি, ২০২৩",
+    "title": "আমার স্ত্রীর জন্য একটি দুর্দান্ত উপহার",
+    "comment": "আমি আমার স্ত্রীর জন্মদিনে এই শাড়ি উপহার দিয়েছিলাম। তিনি এটি পেয়ে খুব খুশি হয়েছেন। প্যাকেজিং এবং ডেলিভারি সার্ভিস খুব ভালো ছিল।"
+  }
+]
